chore: extract RichBodyCell component from TableBody
